End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_modules/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -17,7 +18,9 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -34,8 +37,12 @@ export class MessagesComponent implements OnInit {
 
   deleteMessage(id: number)
   {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex( x => x.id === id), 1);
+    this.confirmService.confirm('Are you sure you want to delete this message?').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex( x => x.id === id), 1);
+        });
+      }
     });
   }
